… docs/deploy/validating-webhook.md
… docs/how-it-works.md
… docs/user-guide/cli-arguments.md
… go.sum
… internal/admission/controller/main_test.go
… internal/admission/controller/server_test.go
… internal/ingress/controller/controller_test.go
… internal/ingress/controller/util.go
… internal/ingress/metric/collectors/controller.go
… internal/ingress/metric/dummy.go
… internal/net/ssl/ssl.go
… internal/net/ssl/ssl_test.go
… vendor/k8s.io/api/admission/v1beta1/doc.go
… vendor/k8s.io/api/admission/v1beta1/generated.pb.go
… vendor/k8s.io/api/admission/v1beta1/generated.proto
… vendor/k8s.io/api/admission/v1beta1/register.go
… vendor/k8s.io/api/admission/v1beta1/types.go
… vendor/k8s.io/api/admission/v1beta1/types_swagger_doc_generated.go
… vendor/k8s.io/api/admission/v1beta1/zz_generated.deepcopy.go