@aledbf 2019-02-22
@tjamet first, thank you for working on this!
That said:
- please rebase
- to update the dependencies please run:
  - rm -rf images/grpc-fortune-teller/ (you can restore that before the commit)
  - make dep-ensure