@aledbf 2019-02-22
@tjamet as you said, we need tests for this :)
What's the behavior when you have multiple deployments with different classes?