@aledbf 2019-02-22
> Regarding multiple deployments, the expected behavior is to accept ingresses that do not [match the class](https://github.com/kubernetes/ingress-nginx/pull/3802/commits/c07f55330b6bb00128ede5a80d4d5c279b6015a9#diff-
86ac9ff9d75a0c5005c116e766a6127dR203).

Yes, I saw that but I am not sure about the webhook behavior with multiple deployments