@ElvinEfendi 2019-02-23
Also fixes https://github.com/kubernetes/ingress-nginx/issues/3588