@tjamet 2019-02-25
@aledbf I am seeing so far 2 parameters that can change the set of ingresses considered for configuration generation in a given controller:
- `--ingress-class`
- `--watch-namespace`
I am hence suspecting multiple deployments to have 4 different ways to consider or not an ingress.

Following your comment, I added the filter based on the namespace name as well for to behave consistently with the deployment. This should handle all combinations of `--ingress-class` and `--watch-namespace`, as soon as the validating webhook is enabled for all deployments.

The approach @ElvinEfendi  is pretty simple in the end, it consists of exposing, outside of the core event loop an HTTP handler, compliant with Kubernetes validating webhook interface, to ensure any ingress that is about to be inserted in the cluster would generate a syntactically valid configuration **_together with the rest of the known ingresses_** using the pre-existing features. This handler should be exposed through a service to a ValidatingWebhookConfiguration object.

As this handler is run outside of the core event loop, I have tested a benchmark that shows the following results (reported by `go test -bench '.' -benchmem -run Benchmark -v` within the `./bin/go-in-docker.sh`):

|number of pre existing ingresses|(go) memory used per request|time per validating request|
|-|-|-|
|0|136342 B/op|25926680 ns/op|
|20|637230 B/op|26366018 ns/op|
|50|1415960 B/op|38304690 ns/op|
|100|2587819 B/op|47551526 ns/op|
|200|5566906 B/op|71309565 ns/op|
|500|12790632 B/op|128302310 ns/op|
|1000|28805841 B/op|226466720 ns/op|

Because the request is served through a service to using the ingress controller as a backend, webhook requests are load-balanced through all ingress controllers and hence reduces the load on a single pod in case multiple ingresses are inserted at the same time.

Another option to get better performances could be to drop the test with all pre-existing ingresses, which could probably be good enough to detect most configuration errors (syntax errors mainly).

Regarding the default enabling of the feature, I would tend to, as a system operator, to recommend to enable it systematically.
Although, it looks to me difficult to enable it by default directly after implementation. This would change the permissions the ingress controller would require to grant it system wide privileges to inject the `ValidatingWebhookConfiguration` object.
Also the way ValidatingWebhook works, they require a self signed certificate generated for the correct `<service>.<namespace>.svc` domain, which would be more difficult to integrate out of the box inside the ingress controller. My approach would rather be to make it easy in the chart to use such a feature to start with. Happy to read other opinions or concerns