@aledbf 2019-03-14
> This could also be integrated afterwards to the helm chart, that I could contribute to as well

Just in case, we also need the plain yaml files, like all the files in https://github.com/kubernetes/ingress-nginx/tree/master/deploy