@tjamet 2019-03-14
@aledbf
I need to add some comments to describe the new methods to handle
certificate rotation, I am planning to do it as soon as I have a stable
internet connection, being on a trip for one week more
I will also update documentation about the certificate generation
Regarding the plain yamls, as it needs the secrets, I added sample
templates, if you have a better idea, I am taker
Happy to update it with anything else required, let me know
If you need to update it meanwhile I have a weak internet connection,
please do


On Thu, 14 Mar 2019 at 18:51, Manuel Alejandro de Brito Fontes <
notifications@github.com> wrote:

> @tjamet <https://github.com/tjamet> is anything I can do to help with
> this PR?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/ingress-nginx/pull/3802#issuecomment-472844140>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AFx9gILBzTGA7rhexMrcSxcy4FUSFaiOks5vWkj1gaJpZM4bL6P->
> .
>