@tjamet 2019-04-04
Hi @zq-david-wang 
> would this validating webhook reject ingress resource that are valid for ingress controller of other kind

The validating webhook is filtering on the `kubernetes.io/ingress.class`, ingress resources that match a different class will be accepted.
In case you run multiple ingress controllers (we do run several nginx ingress controllers) in a cluster, I would expect to find controllers deployed with different classes, hence ingresses for different ingresses kind (traefik in the case you are talking about) will be accepted.

In case both ingresses are run with the same class, both ingress controllers would generate a configuration for the ingress and hence the validating webhook will check the validity of the generated configuration