@aledbf 2019-04-15
@tjamet please rebase and squash the commits