@tjamet 2019-04-15
Rebased and commits squashed